change function into a reuseable component
